backend/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
websockets
python-multipart
